// iris/ui/src/types/api/catalog.gen.ts
export interface components {
    schemas: {
// ...
        AssetTypeInfo: {
            /** @description The name of the catalog where the asset type is */
            catalog_name: string;
            /** @description The name of the schema where the asset type is */
            schema_name: string;
            /** @description The name of the asset type */
            name: string;
            /** @description Full name of the asset type (catalog.schema.name) */
            full_name: string;
            /** @description The comment attached to the asset type */
            comment?: string;
            /** @description List of property definitions for this asset type */
            properties?: components["schemas"]["AssetPropertyDef"][];
            /**
             * Format: int64
             * @description Time at which this asset type was created, in epoch milliseconds
             */
            created_at?: number;
            /** @description The identifier of the user who created the asset type */
            created_by?: string;
            /**
             * Format: int64
             * @description Time at which this asset type was last modified, in epoch milliseconds
             */
            updated_at?: number;
            /** @description The identifier of the user who updated the asset type last time */
            updated_by?: string;
            /** @description The identifier of the user who owns the asset type */
            owner?: string;
        };
        CreateAssetType: {
            /** @description The name of the catalog where the asset type will be created */
            catalog_name: string;
            /** @description The name of the schema where the asset type will be created */
            schema_name: string;
            /** @description The name of the asset type */
            name: string;
            /** @description The comment attached to the asset type */
            comment?: string;
            /** @description List of property definitions for this asset type */
            properties?: components["schemas"]["AssetPropertyDef"][];
        };
        AssetInfo: {
            /** @description The name of the catalog where the asset is */
            catalog_name: string;
            /** @description The name of the schema where the asset is */
            schema_name: string;
            /** @description The name of the asset */
            name: string;
            /** @description Full name of the asset (catalog.schema.name) */
            full_name: string;
            /** @description Full name of the asset type this asset belongs to */
            asset_type_full_name: string;
            /** @description Full name of the parent asset (if this is a child asset) */
            parent_asset_full_name?: string;
            /** @description The comment attached to the asset */
            comment?: string;
            /** @description Key-value properties of the asset */
            properties?: {
                [key: string]: string;
            };
            /** @description List of property bindings for this asset */
            property_bindings?: components["schemas"]["AssetPropertyBinding"][];
            /**
             * Format: int64
             * @description Time at which this asset was created, in epoch milliseconds
             */
            created_at?: number;
            /** @description The identifier of the user who created the asset */
            created_by?: string;
            /**
             * Format: int64
             * @description Time at which this asset was last modified, in epoch milliseconds
             */
            updated_at?: number;
            /** @description The identifier of the user who updated the asset last time */
            updated_by?: string;
            /** @description The identifier of the user who owns the asset */
            owner?: string;
        };
        CreateAsset: {
            /** @description The name of the catalog where the asset will be created */
            catalog_name: string;
            /** @description The name of the schema where the asset will be created */
            schema_name: string;
            /** @description The name of the asset */
            name: string;
            /** @description Full name of the asset type this asset belongs to */
            asset_type_full_name: string;
            /** @description Full name of the parent asset (if this is a child asset) */
            parent_asset_full_name?: string;
            /** @description The comment attached to the asset */
            comment?: string;
            /** @description Key-value properties of the asset */
            properties?: {
                [key: string]: string;
            };
        };
// ...
        AssetList: {
            /** @description List of assets */
            assets?: components["schemas"]["AssetInfo"][];
            /** @description Opaque token to retrieve the next page of results */
            next_page_token?: string;
        };
        AssetTypeList: {
            /** @description List of asset types */
            asset_types?: components["schemas"]["AssetTypeInfo"][];
            /** @description Opaque token to retrieve the next page of results */
            next_page_token?: string;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}